Crate sourceview

Modules

prelude

signal

Structs

Buffer
Implements

Completion
Implements

CompletionActivation

CompletionContext
Implements

CompletionInfo
Implements

CompletionItem
Implements

CompletionProposal
Implements

CompletionProvider
Implements

CompletionWords
Implements

DrawSpacesFlags

Encoding
Feature: v3_14

Error
A generic error capable of representing various error domains (types).

File
Feature: v3_14

FileLoader
Feature: v3_14

FileSaver
Feature: v3_14

FileSaverFlags

Gutter
Implements

GutterRenderer
Implements

GutterRendererPixbuf
Implements

GutterRendererState

GutterRendererText
Implements

Language
Implements

LanguageManager
Implements

Map
Feature: v3_18

Mark
Implements

MarkAttributes
Implements

Object
The base class in the object hierarchy.

PrintCompositor
Implements

Region

RegionIter

SearchContext
Feature: v3_10

SearchSettings
Feature: v3_10

SortFlags

SpaceDrawer
Feature: v3_24

SpaceLocationFlags

SpaceTypeFlags

Style
Implements

StyleScheme
Implements

StyleSchemeChooser
Feature: v3_16

StyleSchemeChooserButton
Feature: v3_16

StyleSchemeChooserWidget
Feature: v3_16

StyleSchemeManager
Implements

Tag
Feature: v3_20

UndoManager
Implements

View
Implements

Enums

BackgroundPatternType

ChangeCaseType

CompressionType

GutterRendererAlignmentMode
The alignment mode of the renderer, when a cell spans multiple lines (due to text wrapping).

NewlineType

SmartHomeEndType

ViewGutterPosition

Traits

BufferExt
Trait containing all Buffer methods.

CompletionContextExt
Trait containing all CompletionContext methods.

CompletionExt
Trait containing all Completion methods.

CompletionInfoExt
Trait containing all CompletionInfo methods.

CompletionItemExt
Trait containing all CompletionItem methods.

CompletionProposalExt
Trait containing all CompletionProposal methods.

CompletionProviderExt
Trait containing all CompletionProvider methods.

CompletionWordsExt
Trait containing all CompletionWords methods.

FileExt
Trait containing all File methods.

FileLoaderExt
Trait containing all FileLoader methods.

FileSaverExt
Trait containing all FileSaver methods.

GutterExt
Trait containing all Gutter methods.

GutterRendererExt
Trait containing all GutterRenderer methods.

GutterRendererPixbufExt
Trait containing all GutterRendererPixbuf methods.

GutterRendererTextExt
Trait containing all GutterRendererText methods.

LanguageExt
Trait containing all Language methods.

LanguageManagerExt
Trait containing all LanguageManager methods.

MapExt
Trait containing all Map methods.

MarkAttributesExt
Trait containing all MarkAttributes methods.

MarkExt
Trait containing all Mark methods.

PrintCompositorExt
Trait containing all PrintCompositor methods.

RegionExt

SearchContextExt
Trait containing all SearchContext methods.

SearchSettingsExt
Trait containing all SearchSettings methods.

SpaceDrawerExt
Trait containing all SpaceDrawer methods.

StyleExt
Trait containing all Style methods.

StyleSchemeChooserExt
Trait containing all StyleSchemeChooser methods.

StyleSchemeExt
Trait containing all StyleScheme methods.

StyleSchemeManagerExt
Trait containing all StyleSchemeManager methods.

TagExt
Trait containing all Tag methods.

UndoManagerExt
Trait containing all UndoManager methods.

ViewExt
Trait containing all View methods.