Trait sourceview::prelude::CompletionContextExt

pub trait CompletionContextExt: 'static {
    fn add_proposals<P: IsA<CompletionProvider>>(
        &self, 
        provider: &P, 
        proposals: &[CompletionProposal], 
        finished: bool
    );

    fn get_activation(&self) -> CompletionActivation;

    fn get_iter(&self) -> Option<TextIter>;

    fn set_property_activation(&self, activation: CompletionActivation);

    fn get_property_completion(&self) -> Option<Completion>;

    fn connect_cancelled<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn emit_cancelled(&self);

    fn connect_property_activation_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all CompletionContext methods.

Implementors

CompletionContext

Required methods

fn add_proposals<P: IsA<CompletionProvider>>(
&self,
provider: &P,
proposals: &[CompletionProposal],
finished: bool
)

Providers can use this function to add proposals to the completion. They can do so asynchronously by means of the finished argument. Providers must ensure that they always call this function with finished set to true once each population (even if no proposals need to be added). Population occurs when the CompletionProvider::populate function is called.

provider

a CompletionProvider.

proposals

The list of proposals to add.

finished

Whether the provider is finished adding proposals.

fn get_activation(&self) -> CompletionActivation

Get the context activation.

Returns

The context activation.

fn get_iter(&self) -> Option<TextIter>

Get the iter at which the completion was invoked. Providers can use this to determine how and if to match proposals.

iter

a gtk::TextIter.

Returns

true if iter is correctly set, false otherwise.

fn set_property_activation(&self, activation: CompletionActivation)

The completion activation

fn get_property_completion(&self) -> Option<Completion>

The Completion associated with the context.

fn connect_cancelled<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

Emitted when the current population of proposals has been cancelled. Providers adding proposals asynchronously should connect to this signal to know when to cancel running proposal queries.

fn emit_cancelled(&self)

fn connect_property_activation_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...