Trait sourceview::prelude::GutterRendererExt

pub trait GutterRendererExt: 'static {
    fn activate(
        &self, 
        iter: &mut TextIter, 
        area: &mut Rectangle, 
        event: &mut Event
    );

    fn begin(
        &self, 
        cr: &Context, 
        background_area: &mut Rectangle, 
        cell_area: &mut Rectangle, 
        start: &mut TextIter, 
        end: &mut TextIter
    );

    fn draw(
        &self, 
        cr: &Context, 
        background_area: &mut Rectangle, 
        cell_area: &mut Rectangle, 
        start: &mut TextIter, 
        end: &mut TextIter, 
        state: GutterRendererState
    );

    fn end(&self);

    fn get_alignment(&self) -> (f32, f32);

    fn get_alignment_mode(&self) -> GutterRendererAlignmentMode;

    fn get_background(&self) -> Option<RGBA>;

    fn get_padding(&self) -> (i32, i32);

    fn get_size(&self) -> i32;

    fn get_view(&self) -> Option<TextView>;

    fn get_visible(&self) -> bool;

    fn get_window_type(&self) -> TextWindowType;

    fn query_activatable(
        &self, 
        iter: &mut TextIter, 
        area: &mut Rectangle, 
        event: &mut Event
    ) -> bool;

    fn query_data(
        &self, 
        start: &mut TextIter, 
        end: &mut TextIter, 
        state: GutterRendererState
    );

    fn query_tooltip(
        &self, 
        iter: &mut TextIter, 
        area: &mut Rectangle, 
        x: i32, 
        y: i32, 
        tooltip: &Tooltip
    ) -> bool;

    fn queue_draw(&self);

    fn set_alignment(&self, xalign: f32, yalign: f32);

    fn set_alignment_mode(&self, mode: GutterRendererAlignmentMode);

    fn set_background(&self, color: Option<&RGBA>);

    fn set_padding(&self, xpad: i32, ypad: i32);

    fn set_size(&self, size: i32);

    fn set_visible(&self, visible: bool);

    fn get_property_background_rgba(&self) -> Option<RGBA>;

    fn set_property_background_rgba(&self, background_rgba: Option<&RGBA>);

    fn get_property_background_set(&self) -> bool;

    fn set_property_background_set(&self, background_set: bool);

    fn get_property_xalign(&self) -> f32;

    fn set_property_xalign(&self, xalign: f32);

    fn get_property_xpad(&self) -> i32;

    fn set_property_xpad(&self, xpad: i32);

    fn get_property_yalign(&self) -> f32;

    fn set_property_yalign(&self, yalign: f32);

    fn get_property_ypad(&self) -> i32;

    fn set_property_ypad(&self, ypad: i32);

    fn connect_activate<F: Fn(&Self, &TextIter, &Rectangle, &Event) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_query_activatable<F: Fn(&Self, &TextIter, &Rectangle, &Event) -> bool + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_query_data<F: Fn(&Self, &TextIter, &TextIter, GutterRendererState) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_query_tooltip<F: Fn(&Self, &TextIter, &Rectangle, i32, i32, &Tooltip) -> bool + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_queue_draw<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_alignment_mode_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_background_rgba_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_background_set_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_size_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_view_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_visible_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_window_type_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_xalign_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_xpad_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_yalign_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_ypad_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all GutterRenderer methods.

Implementors

GutterRendererPixbuf, GutterRendererText, GutterRenderer

Required methods

fn activate(&self, iter: &mut TextIter, area: &mut Rectangle, event: &mut Event)

Emits the GutterRenderer::activate signal of the renderer. This is called from Gutter and should never have to be called manually.

iter

a gtk::TextIter at the start of the line where the renderer is activated

area

a gdk::Rectangle of the cell area where the renderer is activated

event

the event that triggered the activation

fn begin(
&self,
cr: &Context,
background_area: &mut Rectangle,
cell_area: &mut Rectangle,
start: &mut TextIter,
end: &mut TextIter
)

Called when drawing a region begins. The region to be drawn is indicated by start and end. The purpose is to allow the implementation to precompute some state before the draw method is called for each cell.

cr

a cairo::Context

background_area

a gdk::Rectangle

cell_area

a gdk::Rectangle

start

a gtk::TextIter

end

a gtk::TextIter

fn draw(
&self,
cr: &Context,
background_area: &mut Rectangle,
cell_area: &mut Rectangle,
start: &mut TextIter,
end: &mut TextIter,
state: GutterRendererState
)

Main renderering method. Implementations should implement this method to draw onto the cairo context. The background_area indicates the total area of the cell to be drawn. The cell_area indicates the area where content can be drawn (text, images, etc).

The background_area is the cell_area plus the padding on each side (two times the GutterRenderer:xpad horizontally and two times the GutterRenderer:ypad vertically, so that the cell_area is centered inside background_area).

The state argument indicates the current state of the renderer and should be taken into account to properly draw the different possible states (cursor, prelit, selected) if appropriate.

cr

the cairo render context

background_area

a gdk::Rectangle indicating the total area to be drawn

cell_area

a gdk::Rectangle indicating the area to draw content

start

a gtk::TextIter

end

a gtk::TextIter

state

a GutterRendererState

fn end(&self)

Called when drawing a region of lines has ended.

fn get_alignment(&self) -> (f32, f32)

Get the x-alignment and y-alignment of the gutter renderer.

xalign

return location for the x-alignment, or None to ignore.

yalign

return location for the y-alignment, or None to ignore.

fn get_alignment_mode(&self) -> GutterRendererAlignmentMode

Get the alignment mode. The alignment mode describes the manner in which the renderer is aligned (see :xalign and :yalign).

Returns

a GutterRendererAlignmentMode

fn get_background(&self) -> Option<RGBA>

Get the background color of the renderer.

color

return value for a gdk::RGBA

Returns

true if the background color is set, false otherwise

fn get_padding(&self) -> (i32, i32)

Get the x-padding and y-padding of the gutter renderer.

xpad

return location for the x-padding, or None to ignore.

ypad

return location for the y-padding, or None to ignore.

fn get_size(&self) -> i32

Get the size of the renderer.

Returns

the size of the renderer.

fn get_view(&self) -> Option<TextView>

Get the view associated to the gutter renderer

Returns

a gtk::TextView

fn get_visible(&self) -> bool

Get whether the gutter renderer is visible.

Returns

true if the renderer is visible, false otherwise

fn get_window_type(&self) -> TextWindowType

Get the gtk::TextWindowType associated with the gutter renderer.

Returns

a gtk::TextWindowType

fn query_activatable(
&self,
iter: &mut TextIter,
area: &mut Rectangle,
event: &mut Event
) -> bool

Get whether the renderer is activatable at the location in event. This is called from Gutter to determine whether a renderer is activatable using the mouse pointer.

iter

a gtk::TextIter at the start of the line to be activated

area

a gdk::Rectangle of the cell area to be activated

event

the event that triggered the query

Returns

true if the renderer can be activated, false otherwise

fn query_data(
&self,
start: &mut TextIter,
end: &mut TextIter,
state: GutterRendererState
)

Emit the GutterRenderer::query-data signal. This function is called to query for data just before rendering a cell. This is called from the Gutter. Implementations can override the default signal handler or can connect a signal handler externally to the GutterRenderer::query-data signal.

start

a gtk::TextIter.

end

a gtk::TextIter.

state

a GutterRendererState.

fn query_tooltip(
&self,
iter: &mut TextIter,
area: &mut Rectangle,
x: i32,
y: i32,
tooltip: &Tooltip
) -> bool

Emits the GutterRenderer::query-tooltip signal. This function is called from Gutter. Implementations can override the default signal handler or can connect to the signal externally.

iter

a gtk::TextIter.

area

a gdk::Rectangle.

x

The x position of the tooltip.

y

The y position of the tooltip.

tooltip

a gtk::Tooltip.

Returns

true if the tooltip has been set, false otherwise

fn queue_draw(&self)

Emits the GutterRenderer::queue-draw signal of the renderer. Call this from an implementation to inform that the renderer has changed such that it needs to redraw.

fn set_alignment(&self, xalign: f32, yalign: f32)

Set the alignment of the gutter renderer. Both xalign and yalign can be -1, which means the values will not be changed (this allows changing only one of the values).

xalign is the horizontal alignment. Set to 0 for a left alignment. 1 for a right alignment. And 0.5 for centering the cells. yalign is the vertical alignment. Set to 0 for a top alignment. 1 for a bottom alignment.

xalign

the x-alignment

yalign

the y-alignment

fn set_alignment_mode(&self, mode: GutterRendererAlignmentMode)

Set the alignment mode. The alignment mode describes the manner in which the renderer is aligned (see :xalign and :yalign).

mode

a GutterRendererAlignmentMode

fn set_background(&self, color: Option<&RGBA>)

Set the background color of the renderer. If color is set to None, the renderer will not have a background color.

color

a gdk::RGBA or None

fn set_padding(&self, xpad: i32, ypad: i32)

Set the padding of the gutter renderer. Both xpad and ypad can be -1, which means the values will not be changed (this allows changing only one of the values).

xpad is the left and right padding. ypad is the top and bottom padding.

xpad

the x-padding

ypad

the y-padding

fn set_size(&self, size: i32)

Sets the size of the renderer. A value of -1 specifies that the size is to be determined dynamically.

size

the size

fn set_visible(&self, visible: bool)

Set whether the gutter renderer is visible.

visible

the visibility

fn get_property_background_rgba(&self) -> Option<RGBA>

fn set_property_background_rgba(&self, background_rgba: Option<&RGBA>)

fn get_property_background_set(&self) -> bool

fn set_property_background_set(&self, background_set: bool)

fn get_property_xalign(&self) -> f32

The horizontal alignment of the renderer. Set to 0 for a left alignment. 1 for a right alignment. And 0.5 for centering the cells. A value lower than 0 doesn't modify the alignment.

fn set_property_xalign(&self, xalign: f32)

The horizontal alignment of the renderer. Set to 0 for a left alignment. 1 for a right alignment. And 0.5 for centering the cells. A value lower than 0 doesn't modify the alignment.

fn get_property_xpad(&self) -> i32

The left and right padding of the renderer.

fn set_property_xpad(&self, xpad: i32)

The left and right padding of the renderer.

fn get_property_yalign(&self) -> f32

The vertical alignment of the renderer. Set to 0 for a top alignment. 1 for a bottom alignment. And 0.5 for centering the cells. A value lower than 0 doesn't modify the alignment.

fn set_property_yalign(&self, yalign: f32)

The vertical alignment of the renderer. Set to 0 for a top alignment. 1 for a bottom alignment. And 0.5 for centering the cells. A value lower than 0 doesn't modify the alignment.

fn get_property_ypad(&self) -> i32

The top and bottom padding of the renderer.

fn set_property_ypad(&self, ypad: i32)

The top and bottom padding of the renderer.

fn connect_activate<F: Fn(&Self, &TextIter, &Rectangle, &Event) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::activate signal is emitted when the renderer is activated.

iter

a gtk::TextIter

area

a gdk::Rectangle

event

the event that caused the activation

fn connect_query_activatable<F: Fn(&Self, &TextIter, &Rectangle, &Event) -> bool + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::query-activatable signal is emitted when the renderer can possibly be activated.

iter

a gtk::TextIter

area

a gdk::Rectangle

event

the GdkEvent that is causing the activatable query

fn connect_query_data<F: Fn(&Self, &TextIter, &TextIter, GutterRendererState) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::query-data signal is emitted when the renderer needs to be filled with data just before a cell is drawn. This can be used by general renderer implementations to allow render data to be filled in externally.

start

a gtk::TextIter

end

a gtk::TextIter

state

the renderer state

fn connect_query_tooltip<F: Fn(&Self, &TextIter, &Rectangle, i32, i32, &Tooltip) -> bool + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::query-tooltip signal is emitted when the renderer can show a tooltip.

iter

a gtk::TextIter

area

a gdk::Rectangle

x

the x position (in window coordinates)

y

the y position (in window coordinates)

tooltip

a gtk::Tooltip

fn connect_queue_draw<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

The ::queue-draw signal is emitted when the renderer needs to be redrawn. Use GutterRendererExt::queue_draw to emit this signal from an implementation of the GutterRenderer interface.

fn connect_property_alignment_mode_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_background_rgba_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_background_set_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_size_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_view_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_visible_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_window_type_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_xalign_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_xpad_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_yalign_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_ypad_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...