Trait sourceview::prelude::GutterRendererTextExt

pub trait GutterRendererTextExt: 'static {
    fn measure(&self, text: &str) -> (i32, i32);

    fn measure_markup(&self, markup: &str) -> (i32, i32);

    fn set_markup(&self, markup: &str);

    fn set_text(&self, text: &str);

    fn get_property_markup(&self) -> Option<GString>;

    fn get_property_text(&self) -> Option<GString>;

    fn connect_property_markup_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_text_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all GutterRendererText methods.

Implementors

GutterRendererText

Required methods

fn measure(&self, text: &str) -> (i32, i32)

Measures the text provided using the pango layout used by the GutterRendererText.

text

the text to measure.

width

location to store the width of the text in pixels, or None.

height

location to store the height of the text in pixels, or None.

fn measure_markup(&self, markup: &str) -> (i32, i32)

Measures the pango markup provided using the pango layout used by the GutterRendererText.

markup

the pango markup to measure.

width

location to store the width of the text in pixels, or None.

height

location to store the height of the text in pixels, or None.

fn set_markup(&self, markup: &str)

fn set_text(&self, text: &str)

fn get_property_markup(&self) -> Option<GString>

fn get_property_text(&self) -> Option<GString>

fn connect_property_markup_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_text_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...