Trait sourceview::prelude::RegionExt

pub trait RegionExt: 'static {
    fn add_region<P: IsA<Region>>(&self, region_to_add: Option<&P>);

    fn add_subregion(&self, _start: &TextIter, _end: &TextIter);

    fn get_bounds(&self) -> Option<(TextIter, TextIter)>;

    fn get_buffer(&self) -> Option<TextBuffer>;

    fn get_start_region_iter(&self) -> RegionIter;

    fn intersect_region<P: IsA<Region>>(
        &self, 
        region2: Option<&P>
    ) -> Option<Region>;

    fn intersect_subregion(
        &self, 
        _start: &TextIter, 
        _end: &TextIter
    ) -> Option<Region>;

    fn is_empty(&self) -> bool;

    fn subtract_region<P: IsA<Region>>(&self, region_to_subtract: Option<&P>);

    fn subtract_subregion(&self, _start: &TextIter, _end: &TextIter);

    fn to_string(&self) -> GString;
}

Trait containing all Region methods.

Feature: v3_22

Implementors

Region

Required methods

fn add_region<P: IsA<Region>>(&self, region_to_add: Option<&P>)

Adds region_to_add to self. region_to_add is not modified.

Feature: v3_22

region_to_add

the Region to add to self, or None.

fn add_subregion(&self, _start: &TextIter, _end: &TextIter)

Adds the subregion delimited by _start and _end to self.

Feature: v3_22

_start

the start of the subregion.

_end

the end of the subregion.

fn get_bounds(&self) -> Option<(TextIter, TextIter)>

Gets the start and end bounds of the self.

Feature: v3_22

start

iterator to initialize with the start of self, or None.

end

iterator to initialize with the end of self, or None.

Returns

true if start and end have been set successfully (if non-None), or false if the self is empty.

fn get_buffer(&self) -> Option<TextBuffer>

Feature: v3_22

Returns

the gtk::TextBuffer.

fn get_start_region_iter(&self) -> RegionIter

Initializes a RegionIter to the first subregion of self. If self is empty, iter will be initialized to the end iterator.

Feature: v3_22

iter

iterator to initialize to the first subregion.

fn intersect_region<P: IsA<Region>>(
&self,
region2: Option<&P>
) -> Option<Region>

Returns the intersection between self and region2. self and region2 are not modified.

Feature: v3_22

region2

a Region, or None.

Returns

the intersection as a Region object.

fn intersect_subregion(
&self,
_start: &TextIter,
_end: &TextIter
) -> Option<Region>

Returns the intersection between self and the subregion delimited by _start and _end. self is not modified.

Feature: v3_22

_start

the start of the subregion.

_end

the end of the subregion.

Returns

the intersection as a new Region.

fn is_empty(&self) -> bool

Returns whether the self is empty. A None self is considered empty.

Feature: v3_22

Returns

whether the self is empty.

fn subtract_region<P: IsA<Region>>(&self, region_to_subtract: Option<&P>)

Subtracts region_to_subtract from self. region_to_subtract is not modified.

Feature: v3_22

region_to_subtract

the Region to subtract from self, or None.

fn subtract_subregion(&self, _start: &TextIter, _end: &TextIter)

Subtracts the subregion delimited by _start and _end from self.

Feature: v3_22

_start

the start of the subregion.

_end

the end of the subregion.

fn to_string(&self) -> GString

Gets a string represention of self, for debugging purposes.

The returned string contains the character offsets of the subregions. It doesn't include a newline character at the end of the string.

Feature: v3_22

Returns

a string represention of self. Free with g_free when no longer needed.

Loading content...