Trait sourceview::prelude::SpaceDrawerExt

pub trait SpaceDrawerExt: 'static {
    fn bind_matrix_setting<P: IsA<Settings>>(
        &self, 
        settings: &P, 
        key: &str, 
        flags: SettingsBindFlags
    );

    fn get_enable_matrix(&self) -> bool;

    fn get_matrix(&self) -> Option<Variant>;

    fn get_types_for_locations(
        &self, 
        locations: SpaceLocationFlags
    ) -> SpaceTypeFlags;

    fn set_enable_matrix(&self, enable_matrix: bool);

    fn set_matrix(&self, matrix: Option<&Variant>);

    fn set_types_for_locations(
        &self, 
        locations: SpaceLocationFlags, 
        types: SpaceTypeFlags
    );

    fn connect_property_enable_matrix_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_matrix_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all SpaceDrawer methods.

Feature: v3_24

Implementors

SpaceDrawer

Required methods

fn bind_matrix_setting<P: IsA<Settings>>(
&self,
settings: &P,
key: &str,
flags: SettingsBindFlags
)

Binds the SpaceDrawer:matrix property to a gio::Settings key.

The gio::Settings key must be of the same type as the SpaceDrawer:matrix property, that is, "au".

The gio::SettingsExt::bind function cannot be used, because the default GIO mapping functions don't support glib::Variant properties (maybe it will be supported by a future GIO version, in which case this function can be deprecated).

Feature: v3_24

a gio::Settings object.

key

the settings key to bind.

flags

flags for the binding.

fn get_enable_matrix(&self) -> bool

Feature: v3_24

Returns

whether the SpaceDrawer:matrix property is enabled.

fn get_matrix(&self) -> Option<Variant>

Gets the value of the SpaceDrawer:matrix property, as a glib::Variant. An empty array can be returned in case the matrix is a zero matrix.

The SpaceDrawerExt::get_types_for_locations function may be more convenient to use.

Feature: v3_24

Returns

the SpaceDrawer:matrix value as a new floating glib::Variant instance.

fn get_types_for_locations(
&self,
locations: SpaceLocationFlags
) -> SpaceTypeFlags

If only one location is specified, this function returns what kind of white spaces are drawn at that location. The value is retrieved from the SpaceDrawer:matrix property.

If several locations are specified, this function returns the logical AND for those locations. Which means that if a certain kind of white space is present in the return value, then that kind of white space is drawn at all the specified locations.

Feature: v3_24

locations

one or several SpaceLocationFlags.

Returns

a combination of SpaceTypeFlags.

fn set_enable_matrix(&self, enable_matrix: bool)

Sets whether the SpaceDrawer:matrix property is enabled.

Feature: v3_24

enable_matrix

the new value.

fn set_matrix(&self, matrix: Option<&Variant>)

Sets a new value to the SpaceDrawer:matrix property, as a glib::Variant. If matrix is None, then an empty array is set.

If matrix is floating, it is consumed.

The SpaceDrawerExt::set_types_for_locations function may be more convenient to use.

Feature: v3_24

matrix

the new matrix value, or None.

fn set_types_for_locations(
&self,
locations: SpaceLocationFlags,
types: SpaceTypeFlags
)

Modifies the SpaceDrawer:matrix property at the specified locations.

Feature: v3_24

locations

one or several SpaceLocationFlags.

types

a combination of SpaceTypeFlags.

fn connect_property_enable_matrix_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_matrix_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...