Trait sourceview::prelude::StyleSchemeManagerExt

pub trait StyleSchemeManagerExt: 'static {
    fn append_search_path(&self, path: &str);

    fn force_rescan(&self);

    fn get_scheme(&self, scheme_id: &str) -> Option<StyleScheme>;

    fn get_scheme_ids(&self) -> Vec<GString>;

    fn get_search_path(&self) -> Vec<GString>;

    fn prepend_search_path(&self, path: &str);

    fn set_search_path(&self, path: &[&str]);

    fn connect_property_scheme_ids_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_search_path_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all StyleSchemeManager methods.

Implementors

StyleSchemeManager

Required methods

fn append_search_path(&self, path: &str)

Appends path to the list of directories where the self looks for style scheme files. See StyleSchemeManagerExt::set_search_path for details.

path

a directory or a filename.

fn force_rescan(&self)

Mark any currently cached information about the available style scehems as invalid. All the available style schemes will be reloaded next time the self is accessed.

fn get_scheme(&self, scheme_id: &str) -> Option<StyleScheme>

Looks up style scheme by id.

scheme_id

style scheme id to find.

Returns

a StyleScheme object. Returned value is owned by self and must not be unref'ed.

fn get_scheme_ids(&self) -> Vec<GString>

Returns the ids of the available style schemes.

Returns

a None-terminated array of strings containing the ids of the available style schemes or None if no style scheme is available. The array is sorted alphabetically according to the scheme name. The array is owned by the self and must not be modified.

fn get_search_path(&self) -> Vec<GString>

Returns the current search path for the self. See StyleSchemeManagerExt::set_search_path for details.

Returns

a None-terminated array of string containing the search path. The array is owned by the self and must not be modified.

fn prepend_search_path(&self, path: &str)

Prepends path to the list of directories where the self looks for style scheme files. See StyleSchemeManagerExt::set_search_path for details.

path

a directory or a filename.

fn set_search_path(&self, path: &[&str])

Sets the list of directories where the self looks for style scheme files. If path is None, the search path is reset to default.

path

a None-terminated array of strings or None.

fn connect_property_scheme_ids_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_search_path_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...