Struct sourceview::CompletionItem

pub struct CompletionItem(_, _);

Implements

CompletionItemExt, CompletionProposalExt

Methods

impl CompletionItem

fn new<'a, 'b, P: Into<Option<&'a Pixbuf>>, Q: Into<Option<&'b str>>>(
label: &str,
text: &str,
icon: P,
info: Q
) -> CompletionItem

Create a new CompletionItem with label label, icon icon and extra information info. Both icon and info can be None in which case there will be no icon shown and no extra information available.

label

The item label.

text

The item text.

icon

The item icon.

info

The item extra information.

Returns

a new CompletionItem.

fn new_from_stock<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
label: P,
text: &str,
stock: &str,
info: Q
) -> CompletionItem

Creates a new CompletionItem from a stock item. If label is None, the stock label will be used.

Deprecated since 3.10

Use CompletionItem::new instead.

label

The item label.

text

The item text.

stock

The stock icon.

info

The item extra information.

Returns

a new CompletionItem.

fn new_with_markup<'a, 'b, P: Into<Option<&'a Pixbuf>>, Q: Into<Option<&'b str>>>(
markup: &str,
text: &str,
icon: P,
info: Q
) -> CompletionItem

Create a new CompletionItem with markup label markup, icon icon and extra information info. Both icon and info can be None in which case there will be no icon shown and no extra information available.

markup

The item markup label.

text

The item text.

icon

The item icon.

info

The item extra information.

Returns

a new CompletionItem.