Trait sourceview::CompletionExt

pub trait CompletionExt {
    fn add_provider<P: IsA<CompletionProvider>>(
        &self, 
        provider: &P
    ) -> Result<(), Error>;

    fn block_interactive(&self);

    fn get_info_window(&self) -> Option<CompletionInfo>;

    fn get_providers(&self) -> Vec<CompletionProvider>;

    fn get_view(&self) -> Option<View>;

    fn hide(&self);

    fn move_window(&self, iter: &mut TextIter);

    fn remove_provider<P: IsA<CompletionProvider>>(
        &self, 
        provider: &P
    ) -> Result<(), Error>;

    fn show(
        &self, 
        providers: &[CompletionProvider], 
        context: &CompletionContext
    ) -> bool;

    fn unblock_interactive(&self);

    fn get_property_accelerators(&self) -> u32;

    fn set_property_accelerators(&self, accelerators: u32);

    fn get_property_auto_complete_delay(&self) -> u32;

    fn set_property_auto_complete_delay(&self, auto_complete_delay: u32);

    fn get_property_proposal_page_size(&self) -> u32;

    fn set_property_proposal_page_size(&self, proposal_page_size: u32);

    fn get_property_provider_page_size(&self) -> u32;

    fn set_property_provider_page_size(&self, provider_page_size: u32);

    fn get_property_remember_info_visibility(&self) -> bool;

    fn set_property_remember_info_visibility(
        &self, 
        remember_info_visibility: bool
    );

    fn get_property_select_on_show(&self) -> bool;

    fn set_property_select_on_show(&self, select_on_show: bool);

    fn get_property_show_headers(&self) -> bool;

    fn set_property_show_headers(&self, show_headers: bool);

    fn get_property_show_icons(&self) -> bool;

    fn set_property_show_icons(&self, show_icons: bool);

    fn connect_activate_proposal<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn emit_activate_proposal(&self);

    fn connect_hide<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn emit_hide(&self);

    fn connect_move_cursor<F: Fn(&Self, ScrollStep, i32) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn emit_move_cursor(&self, step: ScrollStep, num: i32);

    fn connect_move_page<F: Fn(&Self, ScrollStep, i32) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn emit_move_page(&self, step: ScrollStep, num: i32);

    fn connect_populate_context<F: Fn(&Self, &CompletionContext) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn emit_populate_context(&self, context: &CompletionContext);

    fn connect_show<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn emit_show(&self);

    fn connect_property_accelerators_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_auto_complete_delay_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_proposal_page_size_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_provider_page_size_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_remember_info_visibility_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_select_on_show_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_show_headers_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_show_icons_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_view_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Completion methods.

Implementors

Completion

Required Methods

fn add_provider<P: IsA<CompletionProvider>>(
&self,
provider: &P
) -> Result<(), Error>

Add a new CompletionProvider to the completion object. This will add a reference provider, so make sure to unref your own copy when you no longer need it.

provider

a CompletionProvider.

Returns

true if provider was successfully added, otherwise if error is provided, it will be set with the error and false is returned.

fn block_interactive(&self)

Block interactive completion. This can be used to disable interactive completion when inserting or deleting text from the buffer associated with the completion. Use CompletionExt::unblock_interactive to enable interactive completion again.

This function may be called multiple times. It will continue to block interactive completion until CompletionExt::unblock_interactive has been called the same number of times.

fn get_info_window(&self) -> Option<CompletionInfo>

The info widget is the window where the completion displays optional extra information of the proposal.

Returns

The CompletionInfo window associated with self.

fn get_providers(&self) -> Vec<CompletionProvider>

Get list of providers registered on self. The returned list is owned by the completion and should not be freed.

Returns

list of CompletionProvider.

fn get_view(&self) -> Option<View>

The View associated with self, or None if the view has been destroyed.

Returns

The View associated with self, or None.

fn hide(&self)

Hides the completion if it is active (visible).

fn move_window(&self, iter: &mut TextIter)

Move the completion window to a specific iter.

Deprecated since 3.8

Use CompletionProvider::get_start_iter instead.

iter

a gtk::TextIter.

fn remove_provider<P: IsA<CompletionProvider>>(
&self,
provider: &P
) -> Result<(), Error>

Remove provider from the completion.

provider

a CompletionProvider.

Returns

true if provider was successfully removed, otherwise if error is provided, it will be set with the error and false is returned.

fn show(
&self,
providers: &[CompletionProvider],
context: &CompletionContext
) -> bool

Starts a new completion with the specified CompletionContext and a list of potential candidate providers for completion.

It can be convenient for showing a completion on-the-fly, without the need to add or remove providers to the Completion.

Another solution is to add providers with CompletionExt::add_provider, and implement CompletionProvider::match for each provider.

providers

a list of CompletionProvider, or None.

context

The CompletionContext with which to start the completion.

Returns

true if it was possible to the show completion window.

fn unblock_interactive(&self)

Unblock interactive completion. This can be used after using CompletionExt::block_interactive to enable interactive completion again.

fn get_property_accelerators(&self) -> u32

Number of keyboard accelerators to show for the first proposals. For example, to activate the first proposal, the user can press <keycombo>``<keycap>Alt</keycap>``<keycap>1</keycap>``</keycombo>.

fn set_property_accelerators(&self, accelerators: u32)

Number of keyboard accelerators to show for the first proposals. For example, to activate the first proposal, the user can press <keycombo>``<keycap>Alt</keycap>``<keycap>1</keycap>``</keycombo>.

fn get_property_auto_complete_delay(&self) -> u32

fn set_property_auto_complete_delay(&self, auto_complete_delay: u32)

fn get_property_proposal_page_size(&self) -> u32

fn set_property_proposal_page_size(&self, proposal_page_size: u32)

fn get_property_provider_page_size(&self) -> u32

fn set_property_provider_page_size(&self, provider_page_size: u32)

fn get_property_remember_info_visibility(&self) -> bool

fn set_property_remember_info_visibility(&self, remember_info_visibility: bool)

fn get_property_select_on_show(&self) -> bool

fn set_property_select_on_show(&self, select_on_show: bool)

fn get_property_show_headers(&self) -> bool

fn set_property_show_headers(&self, show_headers: bool)

fn get_property_show_icons(&self) -> bool

fn set_property_show_icons(&self, show_icons: bool)

fn connect_activate_proposal<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

The Completion::activate-proposal signal is a keybinding signal which gets emitted when the user initiates a proposal activation.

Applications should not connect to it, but may emit it with g_signal_emit_by_name if they need to control the proposal activation programmatically.

fn emit_activate_proposal(&self)

fn connect_hide<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

Emitted when the completion window is hidden. The default handler will actually hide the window.

fn emit_hide(&self)

fn connect_move_cursor<F: Fn(&Self, ScrollStep, i32) + 'static>(
&self,
f: F
) -> SignalHandlerId

The Completion::move-cursor signal is a keybinding signal which gets emitted when the user initiates a cursor movement.

The <keycap>Up</keycap>, <keycap>Down</keycap>, <keycap>PageUp</keycap>, <keycap>PageDown</keycap>, <keycap>Home</keycap> and <keycap>End</keycap> keys are bound to the normal behavior expected by those keys.

When step is equal to gtk::ScrollStep::Pages, the page size is defined by the Completion:proposal-page-size property. It is used for the <keycap>PageDown</keycap> and <keycap>PageUp</keycap> keys.

Applications should not connect to it, but may emit it with g_signal_emit_by_name if they need to control the cursor programmatically.

step

The gtk::ScrollStep by which to move the cursor

num

The amount of steps to move the cursor

fn emit_move_cursor(&self, step: ScrollStep, num: i32)

fn connect_move_page<F: Fn(&Self, ScrollStep, i32) + 'static>(
&self,
f: F
) -> SignalHandlerId

The Completion::move-page signal is a keybinding signal which gets emitted when the user initiates a page movement (i.e. switches between provider pages).

<keycombo>``<keycap>Control</keycap>``<keycap>Left</keycap>``</keycombo> is for going to the previous provider. <keycombo>``<keycap>Control</keycap>``<keycap>Right</keycap>``</keycombo> is for going to the next provider. <keycombo>``<keycap>Control</keycap>``<keycap>Home</keycap>``</keycombo> is for displaying all the providers. <keycombo>``<keycap>Control</keycap>``<keycap>End</keycap>``</keycombo> is for going to the last provider.

When step is equal to gtk::ScrollStep::Pages, the page size is defined by the Completion:provider-page-size property.

Applications should not connect to it, but may emit it with g_signal_emit_by_name if they need to control the page selection programmatically.

step

The gtk::ScrollStep by which to move the page

num

The amount of steps to move the page

fn emit_move_page(&self, step: ScrollStep, num: i32)

fn connect_populate_context<F: Fn(&Self, &CompletionContext) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted just before starting to populate the completion with providers. You can use this signal to add additional attributes in the context.

context

The CompletionContext for the current completion

fn emit_populate_context(&self, context: &CompletionContext)

fn connect_show<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

Emitted when the completion window is shown. The default handler will actually show the window.

fn emit_show(&self)

fn connect_property_accelerators_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_auto_complete_delay_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_proposal_page_size_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_provider_page_size_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_remember_info_visibility_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_select_on_show_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_show_headers_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_show_icons_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_view_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId