Trait sourceview::CompletionInfoExt

pub trait CompletionInfoExt {
    fn get_widget(&self) -> Option<Widget>;

    fn set_widget<'a, P: IsA<Widget> + 'a, Q: Into<Option<&'a P>>>(
        &self, 
        widget: Q
    );

    fn connect_before_show<F: Fn(&Self) + 'static>(&self, f: F) -> u64;
}

Trait containing all CompletionInfo methods.

Implementors

CompletionInfo

Required Methods

fn get_widget(&self) -> Option<Widget>

Get the current content widget.

Deprecated since 3.8

Use gtk::Bin::get_child instead.

Returns

The current content widget.

fn set_widget<'a, P: IsA<Widget> + 'a, Q: Into<Option<&'a P>>>(&self, widget: Q)

Sets the content widget of the info window. See that the previous widget will lose a reference and it can be destroyed, so if you do not want this to happen you must use gobject::Object::ref before calling this method.

Deprecated since 3.8

Use gtk::Container::add instead. If there is already a child widget, remove it with gtk::Container::remove.

widget

a gtk::Widget.

fn connect_before_show<F: Fn(&Self) + 'static>(&self, f: F) -> u64

This signal is emitted before any "show" management. You can connect to this signal if you want to change some properties or position before the real "show".

Deprecated since 3.10

This signal should not be used.