Trait sourceview::CompletionItemExt

pub trait CompletionItemExt {
    fn set_gicon<'a, P: IsA<Icon> + 'a, Q: Into<Option<&'a P>>>(&self, gicon: Q);

    fn set_icon<'a, P: Into<Option<&'a Pixbuf>>>(&self, icon: P);

    fn set_icon_name<'a, P: Into<Option<&'a str>>>(&self, icon_name: P);

    fn set_info<'a, P: Into<Option<&'a str>>>(&self, info: P);

    fn set_label<'a, P: Into<Option<&'a str>>>(&self, label: P);

    fn set_markup<'a, P: Into<Option<&'a str>>>(&self, markup: P);

    fn set_text<'a, P: Into<Option<&'a str>>>(&self, text: P);

    fn set_property_gicon<P: IsA<Icon> + IsA<Object> + SetValueOptional>(
        &self, 
        gicon: Option<&P>
    );

    fn set_property_icon(&self, icon: Option<&Pixbuf>);

    fn set_property_icon_name(&self, icon_name: Option<&str>);

    fn set_property_info(&self, info: Option<&str>);

    fn set_property_label(&self, label: Option<&str>);

    fn set_property_markup(&self, markup: Option<&str>);

    fn set_property_text(&self, text: Option<&str>);

    fn connect_property_gicon_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_icon_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_icon_name_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_info_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_label_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_markup_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_text_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all CompletionItem methods.

Implementors

CompletionItem

Required Methods

fn set_gicon<'a, P: IsA<Icon> + 'a, Q: Into<Option<&'a P>>>(&self, gicon: Q)

Feature: v3_24

gicon

the gio::Icon, or None.

fn set_icon<'a, P: Into<Option<&'a Pixbuf>>>(&self, icon: P)

Feature: v3_24

icon

the gdk_pixbuf::Pixbuf, or None.

fn set_icon_name<'a, P: Into<Option<&'a str>>>(&self, icon_name: P)

Feature: v3_24

icon_name

the icon name, or None.

fn set_info<'a, P: Into<Option<&'a str>>>(&self, info: P)

Feature: v3_24

info

the info, or None.

fn set_label<'a, P: Into<Option<&'a str>>>(&self, label: P)

Feature: v3_24

label

the label, or None.

fn set_markup<'a, P: Into<Option<&'a str>>>(&self, markup: P)

Feature: v3_24

markup

the markup, or None.

fn set_text<'a, P: Into<Option<&'a str>>>(&self, text: P)

Feature: v3_24

text

the text, or None.

fn set_property_gicon<P: IsA<Icon> + IsA<Object> + SetValueOptional>(
&self,
gicon: Option<&P>
)

The gio::Icon for the icon to be shown for this proposal.

Feature: v3_18

fn set_property_icon(&self, icon: Option<&Pixbuf>)

The gdk_pixbuf::Pixbuf for the icon to be shown for this proposal.

fn set_property_icon_name(&self, icon_name: Option<&str>)

fn set_property_info(&self, info: Option<&str>)

Optional extra information to be shown for this proposal.

fn set_property_label(&self, label: Option<&str>)

Label to be shown for this proposal.

fn set_property_markup(&self, markup: Option<&str>)

Label with markup to be shown for this proposal.

fn set_property_text(&self, text: Option<&str>)

Proposal text.

fn connect_property_gicon_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_icon_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_icon_name_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_info_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_label_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_markup_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_text_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId