Trait sourceview::FileLoaderExt

pub trait FileLoaderExt {
    fn get_buffer(&self) -> Option<Buffer>;

    fn get_compression_type(&self) -> CompressionType;

    fn get_encoding(&self) -> Option<Encoding>;

    fn get_file(&self) -> Option<File>;

    fn get_location(&self) -> Option<File>;

    fn get_newline_type(&self) -> NewlineType;

    fn set_candidate_encodings(&self, candidate_encodings: &[&Encoding]);

    fn connect_property_buffer_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_file_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_input_stream_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_location_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all FileLoader methods.

Feature: v3_14

Implementors

FileLoader

Required Methods

fn get_buffer(&self) -> Option<Buffer>

Feature: v3_14

Returns

the Buffer to load the contents into.

fn get_compression_type(&self) -> CompressionType

Feature: v3_14

Returns

the detected compression type.

fn get_encoding(&self) -> Option<Encoding>

Feature: v3_14

Returns

the detected file encoding.

fn get_file(&self) -> Option<File>

Feature: v3_14

Returns

the File.

fn get_location(&self) -> Option<File>

Feature: v3_14

Returns

the gio::File to load, or None if an input stream is used.

fn get_newline_type(&self) -> NewlineType

Feature: v3_14

Returns

the detected newline type.

fn set_candidate_encodings(&self, candidate_encodings: &[&Encoding])

Sets the candidate encodings for the file loading. The encodings are tried in the same order as the list.

For convenience, candidate_encodings can contain duplicates. Only the first occurrence of a duplicated encoding is kept in the list.

By default the candidate encodings are (in that order in the list):

If set, the File's encoding as returned by FileExt::get_encoding.

The default candidates as returned by Encoding::get_default_candidates.

Feature: v3_14

candidate_encodings

a list of Encodings.

fn connect_property_buffer_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_file_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_input_stream_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_location_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId