Trait sourceview::LanguageManagerExt

pub trait LanguageManagerExt {
    fn get_language(&self, id: &str) -> Option<Language>;

    fn get_language_ids(&self) -> Vec<String>;

    fn get_search_path(&self) -> Vec<String>;

    fn guess_language<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
        &self, 
        filename: P, 
        content_type: Q
    ) -> Option<Language>;

    fn set_search_path(&self, dirs: &[&str]);

    fn connect_property_language_ids_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_search_path_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all LanguageManager methods.

Implementors

LanguageManager

Required Methods

fn get_language(&self, id: &str) -> Option<Language>

Gets the Language identified by the given id in the language manager.

id

a language id.

Returns

a Language, or None if there is no language identified by the given id. Return value is owned by self and should not be freed.

fn get_language_ids(&self) -> Vec<String>

Returns the ids of the available languages.

Returns

a None-terminated array of strings containing the ids of the available languages or None if no language is available. The array is sorted alphabetically according to the language name. The array is owned by self and must not be modified.

fn get_search_path(&self) -> Vec<String>

Gets the list directories where self looks for language files.

Returns

None-terminated array containg a list of language files directories. The array is owned by self and must not be modified.

fn guess_language<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
&self,
filename: P,
content_type: Q
) -> Option<Language>

Picks a Language for given file name and content type, according to the information in lang files. Either filename or content_type may be None. This function can be used as follows:

<informalexample>``<programlisting> Language *lang; lang = gtk_source_language_manager_guess_language (filename, NULL); gtk_source_buffer_set_language (buffer, lang); </programlisting>``</informalexample>

or

<informalexample>``<programlisting> Language *lang = NULL; gboolean result_uncertain; gchar *content_type;

content_type = g_content_type_guess (filename, NULL, 0, &result_uncertain); if (result_uncertain) { g_free (content_type); content_type = NULL; }

lang = gtk_source_language_manager_guess_language (manager, filename, content_type); gtk_source_buffer_set_language (buffer, lang);

g_free (content_type); </programlisting>``</informalexample>

etc. Use LanguageExt::get_mime_types and LanguageExt::get_globs if you need full control over file -> language mapping.

filename

a filename in Glib filename encoding, or None.

content_type

a content type (as in GIO API), or None.

Returns

a Language, or None if there is no suitable language for given filename and/or content_type. Return value is owned by self and should not be freed.

fn set_search_path(&self, dirs: &[&str])

Sets the list of directories where the self looks for language files. If dirs is None, the search path is reset to default.

<note> <para> At the moment this function can be called only before the language files are loaded for the first time. In practice to set a custom search path for a LanguageManager, you have to call this function right after creating it. </para> </note>

dirs

a None-terminated array of strings or None.

fn connect_property_language_ids_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_search_path_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId