Trait sourceview::MapExt

pub trait MapExt {
    fn get_view(&self) -> Option<View>;

    fn set_view<P: IsA<View>>(&self, view: &P);

    fn get_property_view(&self) -> Option<View>;

    fn set_property_view<P: IsA<View> + IsA<Object> + SetValueOptional>(
        &self, 
        view: Option<&P>
    );

    fn connect_property_view_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Map methods.

Feature: v3_18

Implementors

Map

Required Methods

fn get_view(&self) -> Option<View>

Gets the Map:view property, which is the view this widget is mapping.

Feature: v3_18

Returns

a View or None.

fn set_view<P: IsA<View>>(&self, view: &P)

Sets the view that self will be doing the mapping to.

Feature: v3_18

view

a View

fn get_property_view(&self) -> Option<View>

fn set_property_view<P: IsA<View> + IsA<Object> + SetValueOptional>(
&self,
view: Option<&P>
)

fn connect_property_view_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId