Trait sourceview::MarkExt

pub trait MarkExt: 'static {
    fn get_category(&self) -> Option<GString>;

    fn next(&self, category: Option<&str>) -> Option<Mark>;

    fn prev(&self, category: &str) -> Option<Mark>;
}

Trait containing all Mark methods.

Implementors

Mark

Required methods

fn get_category(&self) -> Option<GString>

Returns the mark category.

Returns

the category of the Mark.

fn next(&self, category: Option<&str>) -> Option<Mark>

Returns the next Mark in the buffer or None if the mark was not added to a buffer. If there is no next mark, None will be returned.

If category is None, looks for marks of any category.

category

a string specifying the mark category, or None.

Returns

the next Mark, or None.

fn prev(&self, category: &str) -> Option<Mark>

Returns the previous Mark in the buffer or None if the mark was not added to a buffer. If there is no previous mark, None is returned.

If category is None, looks for marks of any category

category

a string specifying the mark category, or None.

Returns

the previous Mark, or None.

Loading content...