Trait sourceview::PrintCompositorExt

pub trait PrintCompositorExt {
    fn get_body_font_name(&self) -> Option<String>;

    fn get_bottom_margin(&self, unit: Unit) -> f64;

    fn get_buffer(&self) -> Option<Buffer>;

    fn get_footer_font_name(&self) -> Option<String>;

    fn get_header_font_name(&self) -> Option<String>;

    fn get_highlight_syntax(&self) -> bool;

    fn get_left_margin(&self, unit: Unit) -> f64;

    fn get_line_numbers_font_name(&self) -> Option<String>;

    fn get_n_pages(&self) -> i32;

    fn get_pagination_progress(&self) -> f64;

    fn get_print_footer(&self) -> bool;

    fn get_print_header(&self) -> bool;

    fn get_print_line_numbers(&self) -> u32;

    fn get_right_margin(&self, unit: Unit) -> f64;

    fn get_tab_width(&self) -> u32;

    fn get_top_margin(&self, unit: Unit) -> f64;

    fn get_wrap_mode(&self) -> WrapMode;

    fn set_body_font_name(&self, font_name: &str);

    fn set_bottom_margin(&self, margin: f64, unit: Unit);

    fn set_footer_font_name<'a, P: Into<Option<&'a str>>>(&self, font_name: P);

    fn set_footer_format<'a, 'b, 'c, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>, R: Into<Option<&'c str>>>(
        &self, 
        separator: bool, 
        left: P, 
        center: Q, 
        right: R
    );

    fn set_header_font_name<'a, P: Into<Option<&'a str>>>(&self, font_name: P);

    fn set_header_format<'a, 'b, 'c, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>, R: Into<Option<&'c str>>>(
        &self, 
        separator: bool, 
        left: P, 
        center: Q, 
        right: R
    );

    fn set_highlight_syntax(&self, highlight: bool);

    fn set_left_margin(&self, margin: f64, unit: Unit);

    fn set_line_numbers_font_name<'a, P: Into<Option<&'a str>>>(
        &self, 
        font_name: P
    );

    fn set_print_footer(&self, print: bool);

    fn set_print_header(&self, print: bool);

    fn set_print_line_numbers(&self, interval: u32);

    fn set_right_margin(&self, margin: f64, unit: Unit);

    fn set_tab_width(&self, width: u32);

    fn set_top_margin(&self, margin: f64, unit: Unit);

    fn set_wrap_mode(&self, wrap_mode: WrapMode);

    fn connect_property_body_font_name_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_buffer_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_footer_font_name_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_header_font_name_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_highlight_syntax_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_line_numbers_font_name_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_n_pages_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_print_footer_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_print_header_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_print_line_numbers_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_tab_width_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_wrap_mode_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;
}

Trait containing all PrintCompositor methods.

Implementors

PrintCompositor

Required Methods

fn get_body_font_name(&self) -> Option<String>

Returns the name of the font used to print the text body. The returned string must be freed with g_free.

Feature: v2_2

Returns

a new string containing the name of the font used to print the text body.

fn get_bottom_margin(&self, unit: Unit) -> f64

Gets the bottom margin in units of unit.

Feature: v2_2

unit

the unit for the return value.

Returns

the bottom margin.

fn get_buffer(&self) -> Option<Buffer>

Gets the Buffer associated with the compositor. The returned object reference is owned by the compositor object and should not be unreferenced.

Feature: v2_2

Returns

the Buffer associated with the compositor.

fn get_footer_font_name(&self) -> Option<String>

Returns the name of the font used to print the page footer. The returned string must be freed with g_free.

Feature: v2_2

Returns

a new string containing the name of the font used to print the page footer.

fn get_header_font_name(&self) -> Option<String>

Returns the name of the font used to print the page header. The returned string must be freed with g_free.

Feature: v2_2

Returns

a new string containing the name of the font used to print the page header.

fn get_highlight_syntax(&self) -> bool

Determines whether the printed text will be highlighted according to the buffer rules. Note that highlighting will happen only if the buffer to print has highlighting activated.

Feature: v2_2

Returns

true if the printed output will be highlighted.

fn get_left_margin(&self, unit: Unit) -> f64

Gets the left margin in units of unit.

Feature: v2_2

unit

the unit for the return value.

Returns

the left margin

fn get_line_numbers_font_name(&self) -> Option<String>

Returns the name of the font used to print line numbers on the left margin. The returned string must be freed with g_free.

Feature: v2_2

Returns

a new string containing the name of the font used to print line numbers on the left margin.

fn get_n_pages(&self) -> i32

Returns the number of pages in the document or <code>-1</code> if the document has not been completely paginated.

Feature: v2_2

Returns

the number of pages in the document or <code>-1</code> if the document has not been completely paginated.

fn get_pagination_progress(&self) -> f64

Returns the current fraction of the document pagination that has been completed.

Feature: v2_2

Returns

a fraction from 0.0 to 1.0 inclusive.

fn get_print_footer(&self) -> bool

Determines if a footer is set to be printed for each page. A footer will be printed if this function returns true <emphasis>and</emphasis> some format strings have been specified with PrintCompositorExt::set_footer_format.

Feature: v2_2

Returns

true if the footer is set to be printed.

fn get_print_header(&self) -> bool

Determines if a header is set to be printed for each page. A header will be printed if this function returns true <emphasis>and</emphasis> some format strings have been specified with PrintCompositorExt::set_header_format.

Feature: v2_2

Returns

true if the header is set to be printed.

fn get_print_line_numbers(&self) -> u32

Returns the interval used for line number printing. If the value is 0, no line numbers will be printed. The default value is 1 (i.e. numbers printed in all lines).

Feature: v2_2

Returns

the interval of printed line numbers.

fn get_right_margin(&self, unit: Unit) -> f64

Gets the right margin in units of unit.

Feature: v2_2

unit

the unit for the return value.

Returns

the right margin.

fn get_tab_width(&self) -> u32

Returns the width of tabulation in characters for printed text.

Feature: v2_2

Returns

width of tab.

fn get_top_margin(&self, unit: Unit) -> f64

Gets the top margin in units of unit.

Feature: v2_2

unit

the unit for the return value.

Returns

the top margin.

fn get_wrap_mode(&self) -> WrapMode

Gets the line wrapping mode for the printed text.

Feature: v2_2

Returns

the line wrap mode.

fn set_body_font_name(&self, font_name: &str)

Sets the default font for the printed text.

font_name should be a string representation of a font description Pango can understand. (e.g. "Monospace 10"). See pango::FontDescription::from_string for a description of the format of the string representation.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

font_name

the name of the default font for the body text.

fn set_bottom_margin(&self, margin: f64, unit: Unit)

Sets the bottom margin used by self.

Feature: v2_2

margin

the new bottom margin in units of unit.

unit

the units for margin.

fn set_footer_font_name<'a, P: Into<Option<&'a str>>>(&self, font_name: P)

Sets the font for printing the page footer. If None is supplied, the default font (i.e. the one being used for the text) will be used instead.

font_name should be a string representation of a font description Pango can understand. (e.g. "Monospace 10"). See pango::FontDescription::from_string for a description of the format of the string representation.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

font_name

the name of the font for the footer text, or None.

fn set_footer_format<'a, 'b, 'c, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>, R: Into<Option<&'c str>>>(
&self,
separator: bool,
left: P,
center: Q,
right: R
)

See PrintCompositorExt::set_header_format for more information about the parameters.

Feature: v2_2

separator

true if you want a separator line to be printed.

left

a format string to print on the left of the footer.

center

a format string to print on the center of the footer.

right

a format string to print on the right of the footer.

fn set_header_font_name<'a, P: Into<Option<&'a str>>>(&self, font_name: P)

Sets the font for printing the page header. If None is supplied, the default font (i.e. the one being used for the text) will be used instead.

font_name should be a string representation of a font description Pango can understand. (e.g. "Monospace 10"). See pango::FontDescription::from_string for a description of the format of the string representation.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

font_name

the name of the font for header text, or None.

fn set_header_format<'a, 'b, 'c, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>, R: Into<Option<&'c str>>>(
&self,
separator: bool,
left: P,
center: Q,
right: R
)

Sets strftime like header format strings, to be printed on the left, center and right of the top of each page. The strings may include strftime(3) codes which will be expanded at print time. A subset of strftime codes are accepted, see glib::DateTime::format for more details on the accepted format specifiers. Additionally the following format specifiers are accepted: - N: the page number - Q: the page count.

separator specifies if a solid line should be drawn to separate the header from the document text.

If None is given for any of the three arguments, that particular string will not be printed.

For the header to be printed, in addition to specifying format strings, you need to enable header printing with PrintCompositorExt::set_print_header.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

separator

true if you want a separator line to be printed.

left

a format string to print on the left of the header.

center

a format string to print on the center of the header.

right

a format string to print on the right of the header.

fn set_highlight_syntax(&self, highlight: bool)

Sets whether the printed text will be highlighted according to the buffer rules. Both color and font style are applied.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

highlight

whether syntax should be highlighted.

fn set_left_margin(&self, margin: f64, unit: Unit)

Sets the left margin used by self.

Feature: v2_2

margin

the new left margin in units of unit.

unit

the units for margin.

fn set_line_numbers_font_name<'a, P: Into<Option<&'a str>>>(&self, font_name: P)

Sets the font for printing line numbers on the left margin. If None is supplied, the default font (i.e. the one being used for the text) will be used instead.

font_name should be a string representation of a font description Pango can understand. (e.g. "Monospace 10"). See pango::FontDescription::from_string for a description of the format of the string representation.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

font_name

the name of the font for line numbers, or None.

fn set_print_footer(&self, print: bool)

Sets whether you want to print a footer in each page. The footer consists of three pieces of text and an optional line separator, configurable with PrintCompositorExt::set_footer_format.

Note that by default the footer format is unspecified, and if it's empty it will not be printed, regardless of this setting.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

print

true if you want the footer to be printed.

fn set_print_header(&self, print: bool)

Sets whether you want to print a header in each page. The header consists of three pieces of text and an optional line separator, configurable with PrintCompositorExt::set_header_format.

Note that by default the header format is unspecified, and if it's empty it will not be printed, regardless of this setting.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

print

true if you want the header to be printed.

fn set_print_line_numbers(&self, interval: u32)

Sets the interval for printed line numbers. If interval is 0 no numbers will be printed. If greater than 0, a number will be printed every interval lines (i.e. 1 will print all line numbers).

Maximum accepted value for interval is 100.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

interval

interval for printed line numbers.

fn set_right_margin(&self, margin: f64, unit: Unit)

Sets the right margin used by self.

Feature: v2_2

margin

the new right margin in units of unit.

unit

the units for margin.

fn set_tab_width(&self, width: u32)

Sets the width of tabulation in characters for printed text.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

width

width of tab in characters.

fn set_top_margin(&self, margin: f64, unit: Unit)

Sets the top margin used by self.

Feature: v2_2

margin

the new top margin in units of unit

unit

the units for margin

fn set_wrap_mode(&self, wrap_mode: WrapMode)

Sets the line wrapping mode for the printed text.

This function cannot be called anymore after the first call to the PrintCompositorExt::paginate function.

Feature: v2_2

wrap_mode

a gtk::WrapMode.

fn connect_property_body_font_name_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64

fn connect_property_buffer_notify<F: Fn(&Self) + 'static>(&self, f: F) -> u64

fn connect_property_footer_font_name_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64

fn connect_property_header_font_name_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64

fn connect_property_highlight_syntax_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64

fn connect_property_line_numbers_font_name_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64

fn connect_property_n_pages_notify<F: Fn(&Self) + 'static>(&self, f: F) -> u64

fn connect_property_print_footer_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64

fn connect_property_print_header_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64

fn connect_property_print_line_numbers_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64

fn connect_property_tab_width_notify<F: Fn(&Self) + 'static>(&self, f: F) -> u64

fn connect_property_wrap_mode_notify<F: Fn(&Self) + 'static>(&self, f: F) -> u64