Trait sourceview::SearchContextExt

pub trait SearchContextExt {
    fn backward(&self, iter: &TextIter) -> Option<(TextIter, TextIter)>;
    fn forward(&self, iter: &TextIter) -> Option<(TextIter, TextIter)>;
    fn get_buffer(&self) -> Option<Buffer>;
    fn get_highlight(&self) -> bool;
    fn get_match_style(&self) -> Option<Style>;
    fn get_occurrence_position(
        &self, 
        match_start: &TextIter, 
        match_end: &TextIter
    ) -> i32;
    fn get_occurrences_count(&self) -> i32;
    fn get_regex_error(&self) -> Option<Error>;
    fn get_settings(&self) -> Option<SearchSettings>;
    fn replace(
        &self, 
        match_start: &TextIter, 
        match_end: &TextIter, 
        replace: &str, 
        replace_length: i32
    ) -> Result<(), Error>;
    fn replace_all(
        &self, 
        replace: &str, 
        replace_length: i32
    ) -> Result<(), Error>;
    fn set_highlight(&self, highlight: bool);
    fn set_match_style<'a, P: Into<Option<&'a Style>>>(&self, match_style: P);
    fn set_settings<'a, P: Into<Option<&'a SearchSettings>>>(&self, settings: P);
}

Trait containing all SearchContext methods.

Feature: v3_10

Implementors

SearchContext

Required Methods

fn backward(&self, iter: &TextIter) -> Option<(TextIter, TextIter)>

Synchronous backward search. It is recommended to use the asynchronous functions instead, to not block the user interface. However, if you are sure that the buffer is small, this function is more convenient to use.

Feature: v3_10

iter

start of search.

match_start

return location for start of match, or None.

match_end

return location for end of match, or None.

Returns

whether a match was found.

fn forward(&self, iter: &TextIter) -> Option<(TextIter, TextIter)>

Synchronous forward search. It is recommended to use the asynchronous functions instead, to not block the user interface. However, if you are sure that the buffer is small, this function is more convenient to use.

Feature: v3_10

iter

start of search.

match_start

return location for start of match, or None.

match_end

return location for end of match, or None.

Returns

whether a match was found.

fn get_buffer(&self) -> Option<Buffer>

Feature: v3_10

Returns

the associated buffer.

fn get_highlight(&self) -> bool

Feature: v3_10

Returns

whether to highlight the search occurrences.

fn get_match_style(&self) -> Option<Style>

Feature: v3_16

Returns

the Style to apply on search matches.

fn get_occurrence_position(
&self,
match_start: &TextIter,
match_end: &TextIter
) -> i32

Gets the position of a search occurrence. If the buffer is not already fully scanned, the position may be unknown, and -1 is returned. If 0 is returned, it means that this part of the buffer has already been scanned, and that match_start and match_end don't delimit an occurrence.

Feature: v3_10

match_start

the start of the occurrence.

match_end

the end of the occurrence.

Returns

the position of the search occurrence. The first occurrence has the position 1 (not 0). Returns 0 if match_start and match_end don't delimit an occurrence. Returns -1 if the position is not yet known.

fn get_occurrences_count(&self) -> i32

Gets the total number of search occurrences. If the buffer is not already fully scanned, the total number of occurrences is unknown, and -1 is returned.

Feature: v3_10

Returns

the total number of search occurrences, or -1 if unknown.

fn get_regex_error(&self) -> Option<Error>

Regular expression patterns must follow certain rules. If SearchSettings:search-text breaks a rule, the error can be retrieved with this function. The error domain is G_REGEX_ERROR.

Free the return value with glib::Error::free.

Feature: v3_10

Returns

the glib::Error, or None if the pattern is valid.

fn get_settings(&self) -> Option<SearchSettings>

Feature: v3_10

Returns

the search settings.

fn replace(
&self,
match_start: &TextIter,
match_end: &TextIter,
replace: &str,
replace_length: i32
) -> Result<(), Error>

Replaces a search match by another text. If match_start and match_end doesn't correspond to a search match, false is returned.

For a regular expression replacement, you can check if replace is valid by calling glib::Regex::check_replacement. The replace text can contain backreferences; read the glib::Regex::replace documentation for more details.

Feature: v3_10

match_start

the start of the match to replace.

match_end

the end of the match to replace.

replace

the replacement text.

replace_length

the length of replace in bytes, or -1.

Returns

whether the match has been replaced.

fn replace_all(&self, replace: &str, replace_length: i32) -> Result<(), Error>

Replaces all search matches by another text. It is a synchronous function, so it can block the user interface.

For a regular expression replacement, you can check if replace is valid by calling glib::Regex::check_replacement. The replace text can contain backreferences; read the glib::Regex::replace documentation for more details.

Feature: v3_10

replace

the replacement text.

replace_length

the length of replace in bytes, or -1.

Returns

the number of replaced matches.

fn set_highlight(&self, highlight: bool)

Enables or disables the search occurrences highlighting.

Feature: v3_10

highlight

the setting.

fn set_match_style<'a, P: Into<Option<&'a Style>>>(&self, match_style: P)

Set the style to apply on search matches. If match_style is None, default theme's scheme 'match-style' will be used. To enable or disable the search highlighting, use SearchContextExt::set_highlight.

Feature: v3_16

match_style

a Style.

fn set_settings<'a, P: Into<Option<&'a SearchSettings>>>(&self, settings: P)

Associate a SearchSettings with the search context. If settings is None, a new one will be created.

The search context holds a reference to settings.

Feature: v3_10

the new SearchSettings, or None.