Trait sourceview::SearchSettingsExt

pub trait SearchSettingsExt: 'static {
    fn get_at_word_boundaries(&self) -> bool;

    fn get_case_sensitive(&self) -> bool;

    fn get_regex_enabled(&self) -> bool;

    fn get_search_text(&self) -> Option<GString>;

    fn get_wrap_around(&self) -> bool;

    fn set_at_word_boundaries(&self, at_word_boundaries: bool);

    fn set_case_sensitive(&self, case_sensitive: bool);

    fn set_regex_enabled(&self, regex_enabled: bool);

    fn set_search_text(&self, search_text: Option<&str>);

    fn set_wrap_around(&self, wrap_around: bool);

    fn connect_property_at_word_boundaries_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_case_sensitive_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_regex_enabled_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_search_text_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_wrap_around_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all SearchSettings methods.

Feature: v3_10

Implementors

SearchSettings

Required methods

fn get_at_word_boundaries(&self) -> bool

Feature: v3_10

Returns

whether to search at word boundaries.

fn get_case_sensitive(&self) -> bool

Feature: v3_10

Returns

whether the search is case sensitive.

fn get_regex_enabled(&self) -> bool

Feature: v3_10

Returns

whether to search by regular expressions.

fn get_search_text(&self) -> Option<GString>

Gets the text to search. The return value must not be freed.

You may be interested to call gtk_source_utils_escape_search_text after this function.

Feature: v3_10

Returns

the text to search, or None if the search is disabled.

fn get_wrap_around(&self) -> bool

Feature: v3_10

Returns

whether to wrap around the search.

fn set_at_word_boundaries(&self, at_word_boundaries: bool)

Change whether the search is done at word boundaries. If at_word_boundaries is true, a search match must start and end a word. The match can span multiple words. See also gtk::TextIter::starts_word and gtk::TextIter::ends_word.

Feature: v3_10

at_word_boundaries

the setting.

fn set_case_sensitive(&self, case_sensitive: bool)

Enables or disables the case sensitivity for the search.

Feature: v3_10

case_sensitive

the setting.

fn set_regex_enabled(&self, regex_enabled: bool)

Enables or disables whether to search by regular expressions. If enabled, the SearchSettings:search-text property contains the pattern of the regular expression.

SearchContext uses glib::Regex when regex search is enabled. See the Regular expression syntax page in the GLib reference manual.

Feature: v3_10

regex_enabled

the setting.

fn set_search_text(&self, search_text: Option<&str>)

Sets the text to search. If search_text is None or is empty, the search will be disabled. A copy of search_text will be made, so you can safely free search_text after a call to this function.

You may be interested to call gtk_source_utils_unescape_search_text before this function.

Feature: v3_10

search_text

the nul-terminated text to search, or None to disable the search.

fn set_wrap_around(&self, wrap_around: bool)

Enables or disables the wrap around search. If wrap_around is true, the forward search continues at the beginning of the buffer if no search occurrences are found. Similarly, the backward search continues to search at the end of the buffer.

Feature: v3_10

wrap_around

the setting.

fn connect_property_at_word_boundaries_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_case_sensitive_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_regex_enabled_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_search_text_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_wrap_around_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...