Trait sourceview::UndoManagerExt

pub trait UndoManagerExt: 'static {
    fn begin_not_undoable_action(&self);

    fn can_redo(&self) -> bool;

    fn can_redo_changed(&self);

    fn can_undo(&self) -> bool;

    fn can_undo_changed(&self);

    fn end_not_undoable_action(&self);

    fn redo(&self);

    fn undo(&self);

    fn connect_can_redo_changed<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn emit_can_redo_changed(&self);

    fn connect_can_undo_changed<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn emit_can_undo_changed(&self);
}

Trait containing all UndoManager methods.

Implementors

UndoManager

Required methods

fn begin_not_undoable_action(&self)

Begin a not undoable action on the buffer. All changes between this call and the call to UndoManager::end_not_undoable_action cannot be undone. This function should be re-entrant.

fn can_redo(&self) -> bool

Get whether there are redo operations available.

Returns

true if there are redo operations available, false otherwise

fn can_redo_changed(&self)

Emits the UndoManager::can-redo-changed signal.

fn can_undo(&self) -> bool

Get whether there are undo operations available.

Returns

true if there are undo operations available, false otherwise

fn can_undo_changed(&self)

Emits the UndoManager::can-undo-changed signal.

fn end_not_undoable_action(&self)

Ends a not undoable action on the buffer.

fn redo(&self)

Perform a single redo. Calling this function when there are no redo operations available is an error. Use UndoManager::can_redo to find out if there are redo operations available.

fn undo(&self)

Perform a single undo. Calling this function when there are no undo operations available is an error. Use UndoManager::can_undo to find out if there are undo operations available.

fn connect_can_redo_changed<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when the ability to redo has changed.

fn emit_can_redo_changed(&self)

fn connect_can_undo_changed<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when the ability to undo has changed.

fn emit_can_undo_changed(&self)

Loading content...