Struct sourceview::Buffer

pub struct Buffer(_, _);

Implements

BufferExt, gtk::TextBufferExt

Implementations

impl Buffer

pub fn new<P: IsA<TextTagTable>>(table: Option<&P>) -> Buffer

Creates a new source buffer.

table

a gtk::TextTagTable, or None to create a new one.

Returns

a new source buffer.

pub fn new_with_language<P: IsA<Language>>(language: &P) -> Buffer

Creates a new source buffer using the highlighting patterns in language. This is equivalent to creating a new source buffer with a new tag table and then calling BufferExt::set_language.

language

a Language.

Returns

a new source buffer which will highlight text according to the highlighting patterns in language.