Struct sourceview::Encoding

pub struct Encoding(_);

Feature: v3_14

Methods

impl Encoding

pub fn get_charset(&self) -> Option<String>

Gets the character set of the Encoding, such as "UTF-8" or "ISO-8859-1".

Feature: v3_14

Returns

the character set of the Encoding.

pub fn get_name(&self) -> Option<String>

Gets the name of the Encoding such as "Unicode" or "Western".

Feature: v3_14

Returns

the name of the Encoding.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_all() -> Vec<Encoding>

Gets all encodings.

Feature: v3_14

Returns

a list of all Encoding's. Free with glib::SList::free.

pub fn get_current() -> Option<Encoding>

Gets the Encoding for the current locale. See also g_get_charset.

Feature: v3_14

Returns

the current locale encoding.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_default_candidates() -> Vec<Encoding>

Gets the list of default candidate encodings to try when loading a file. See FileLoaderExt::set_candidate_encodings.

This function returns a different list depending on the current locale (i.e. language, country and default encoding). The UTF-8 encoding and the current locale encoding are guaranteed to be present in the returned list.

Feature: v3_18

Returns

the list of default candidate encodings. Free with glib::SList::free.

pub fn get_from_charset(charset: &str) -> Option<Encoding>

Gets a Encoding from a character set such as "UTF-8" or "ISO-8859-1".

Feature: v3_14

charset

a character set.

Returns

the corresponding Encoding, or None if not found.

pub fn get_utf8() -> Option<Encoding>

Feature: v3_14

Returns

the UTF-8 encoding.