Trait sourceview::GutterExt

pub trait GutterExt: 'static {
    fn get_renderer_at_pos(&self, x: i32, y: i32) -> Option<GutterRenderer>;

    fn get_view(&self) -> Option<View>;

    fn get_window(&self) -> Option<Window>;

    fn get_window_type(&self) -> TextWindowType;

    fn insert<P: IsA<GutterRenderer>>(
        &self, 
        renderer: &P, 
        position: i32
    ) -> bool;

    fn queue_draw(&self);

    fn remove<P: IsA<GutterRenderer>>(&self, renderer: &P);

    fn reorder<P: IsA<GutterRenderer>>(&self, renderer: &P, position: i32);

    fn set_padding(&self, xpad: i32, ypad: i32);

    fn get_property_view(&self) -> Option<View>;

    fn get_property_window_type(&self) -> TextWindowType;

    fn get_property_xpad(&self) -> i32;

    fn set_property_xpad(&self, xpad: i32);

    fn get_property_ypad(&self) -> i32;

    fn set_property_ypad(&self, ypad: i32);

    fn connect_property_xpad_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_ypad_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Gutter methods.

Implementors

Gutter

Required methods

fn get_renderer_at_pos(&self, x: i32, y: i32) -> Option<GutterRenderer>

Finds the GutterRenderer at (x, y).

x

The x position to get identified.

y

The y position to get identified.

Returns

the renderer at (x, y) or None.

fn get_view(&self) -> Option<View>

Feature: v3_24

Returns

the associated View.

fn get_window(&self) -> Option<Window>

Get the gdk::Window of the gutter. The window will only be available when the gutter has at least one, non-zero width, cell renderer packed.

Deprecated since 3.12

Use gtk::TextViewExt::get_window instead.

Returns

the gdk::Window of the gutter, or None if the gutter has no window.

fn get_window_type(&self) -> TextWindowType

Feature: v3_24

Returns

the gtk::TextWindowType of self.

fn insert<P: IsA<GutterRenderer>>(&self, renderer: &P, position: i32) -> bool

Insert renderer into the gutter. If renderer is yet unowned then gutter claims its ownership. Otherwise just increases renderer's reference count. renderer cannot be already inserted to another gutter.

renderer

a gutter renderer (must inherit from GutterRenderer).

position

the renderer position.

Returns

true if operation succeeded. Otherwise false.

fn queue_draw(&self)

Invalidates the drawable area of the gutter. You can use this to force a redraw of the gutter if something has changed and needs to be redrawn.

fn remove<P: IsA<GutterRenderer>>(&self, renderer: &P)

Removes renderer from self.

renderer

a GutterRenderer.

fn reorder<P: IsA<GutterRenderer>>(&self, renderer: &P, position: i32)

Reorders renderer in self to new position.

renderer

a gtk::CellRenderer.

position

the new renderer position.

fn set_padding(&self, xpad: i32, ypad: i32)

Deprecated since 3.12

Use GutterRendererExt::set_padding instead.

fn get_property_view(&self) -> Option<View>

The View of the gutter.

fn get_property_window_type(&self) -> TextWindowType

The text window type on which the window is placed.

fn get_property_xpad(&self) -> i32

The x-padding.

Deprecated since 3.12

Use the GutterRenderer's GutterRenderer:xpad property instead.

fn set_property_xpad(&self, xpad: i32)

The x-padding.

Deprecated since 3.12

Use the GutterRenderer's GutterRenderer:xpad property instead.

fn get_property_ypad(&self) -> i32

The y-padding.

Deprecated since 3.12

Use the GutterRenderer's GutterRenderer:ypad property instead.

fn set_property_ypad(&self, ypad: i32)

The y-padding.

Deprecated since 3.12

Use the GutterRenderer's GutterRenderer:ypad property instead.

fn connect_property_xpad_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_ypad_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...