Trait sourceview::MarkAttributesExt

pub trait MarkAttributesExt: 'static {
    fn get_background(&self) -> Option<RGBA>;

    fn get_gicon(&self) -> Option<Icon>;

    fn get_icon_name(&self) -> Option<GString>;

    fn get_stock_id(&self) -> Option<GString>;

    fn get_tooltip_markup<P: IsA<Mark>>(&self, mark: &P) -> Option<GString>;

    fn get_tooltip_text<P: IsA<Mark>>(&self, mark: &P) -> Option<GString>;

    fn set_background(&self, background: &RGBA);

    fn set_gicon<P: IsA<Icon>>(&self, gicon: &P);

    fn set_icon_name(&self, icon_name: &str);

    fn set_pixbuf(&self, pixbuf: &Pixbuf);

    fn set_stock_id(&self, stock_id: &str);

    fn get_property_pixbuf(&self) -> Option<Pixbuf>;

    fn connect_query_tooltip_markup<F: Fn(&Self, &Mark) -> String + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_query_tooltip_text<F: Fn(&Self, &Mark) -> String + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_background_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_gicon_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_icon_name_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_pixbuf_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_stock_id_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all MarkAttributes methods.

Implementors

MarkAttributes

Required methods

fn get_background(&self) -> Option<RGBA>

Stores background color in background.

background

a gdk::RGBA.

Returns

whether background color for self was set.

fn get_gicon(&self) -> Option<Icon>

Gets a gio::Icon to be used as a base for rendered icon. Note that the icon can be None if it wasn't set earlier.

Returns

An icon. The icon belongs to self and should not be unreffed.

fn get_icon_name(&self) -> Option<GString>

Gets a name of an icon to be used as a base for rendered icon. Note that the icon name can be None if it wasn't set earlier.

Returns

An icon name. The string belongs to self and should not be freed.

fn get_stock_id(&self) -> Option<GString>

Gets a stock id of an icon used by this attributes. Note that the stock id can be None if it wasn't set earlier.

Deprecated since 3.10

Don't use this function.

Returns

Stock id. Returned string is owned by self and shouldn't be freed.

fn get_tooltip_markup<P: IsA<Mark>>(&self, mark: &P) -> Option<GString>

Queries for a tooltip by emitting a MarkAttributes::query-tooltip-markup signal. The tooltip may contain a markup.

mark

a Mark.

Returns

A tooltip. The returned string should be freed by using g_free when done with it.

fn get_tooltip_text<P: IsA<Mark>>(&self, mark: &P) -> Option<GString>

Queries for a tooltip by emitting a MarkAttributes::query-tooltip-text signal. The tooltip is a plain text.

mark

a Mark.

Returns

A tooltip. The returned string should be freed by using g_free when done with it.

fn set_background(&self, background: &RGBA)

Sets background color to the one given in background.

background

a gdk::RGBA.

fn set_gicon<P: IsA<Icon>>(&self, gicon: &P)

Sets an icon to be used as a base for rendered icon.

gicon

a gio::Icon to be used.

fn set_icon_name(&self, icon_name: &str)

Sets a name of an icon to be used as a base for rendered icon.

icon_name

name of an icon to be used.

fn set_pixbuf(&self, pixbuf: &Pixbuf)

Sets a pixbuf to be used as a base for rendered icon.

pixbuf

a gdk_pixbuf::Pixbuf to be used.

fn set_stock_id(&self, stock_id: &str)

Sets stock id to be used as a base for rendered icon.

Deprecated since 3.10

Don't use this function.

stock_id

a stock id.

fn get_property_pixbuf(&self) -> Option<Pixbuf>

A gdk_pixbuf::Pixbuf that may be a base of a rendered icon.

fn connect_query_tooltip_markup<F: Fn(&Self, &Mark) -> String + 'static>(
&self,
f: F
) -> SignalHandlerId

The code should connect to this signal to provide a tooltip for given mark. The tooltip can contain a markup.

mark

The Mark.

Returns

A tooltip. The string should be freed with g_free when done with it.

fn connect_query_tooltip_text<F: Fn(&Self, &Mark) -> String + 'static>(
&self,
f: F
) -> SignalHandlerId

The code should connect to this signal to provide a tooltip for given mark. The tooltip should be just a plain text.

mark

The Mark.

Returns

A tooltip. The string should be freed with g_free when done with it.

fn connect_property_background_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_gicon_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_icon_name_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_pixbuf_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_stock_id_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...